+++ packages/mdx/src/remark/index.ts
@@ -1,6 +1,6 @@
 import type { PluggableList } from 'unified'
 
-import remarkAddFrontMatter from './add-front-matter.js'
+import remarkAddFrontmatter from './add-frontmatter.js'
 import remarkAddSections from './add-sections.js'
 import remarkGfm from './gfm.js'
 import remarkRemoveImmediateParagraphs from './remove-immediate-paragraphs.js'
@@ -9,7 +9,7 @@ import remarkTransformRelativeLinks from './transform-relative-links.js'
 import remarkTypography from './typography.js'
 
 export const remarkPlugins: PluggableList = [
-  remarkAddFrontMatter,
+  remarkAddFrontmatter,
   remarkAddSections,
   remarkGfm,
   remarkRemoveImmediateParagraphs,